requirements.txt
lxml==4.6.3
pandas==1.2.4
snowflake-connector-python==2.4.5
apache-airflow-providers-snowflake==1.3.0
snowflake-sqlalchemy==1.2.4
apache-airflow-providers-amazon==1.4.0
requests==2.25.1
urllib3==1.25.11
openpyxl==3.0.7